requirements.txt
pendulum==3.0.0
apache-airflow==apache-airflow
requests==2.32.3
google-auth==2.35.0
google-cloud-secret-manager==2.21.0
google-api-core==2.11.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-api-python-client==2.36.0
google-cloud-core==2.3.2
